+++ cli/dstack/api/hub/__init__.py
@@ -1,0 +1,1 @@
+from dstack.api.hub._client import HubClient
