+++ cli/dstack/backend/base/runs.py
@@ -2,28 +2,19 @@ from typing import List
 
 import yaml
 
-from dstack.backend.base import BackendType, jobs, runners
+from dstack.backend.base import jobs, runners
 from dstack.backend.base.compute import Compute
 from dstack.backend.base.storage import Storage
 from dstack.core.app import AppHead
 from dstack.core.artifact import ArtifactHead
 from dstack.core.job import JobErrorCode, JobHead, JobStatus
-from dstack.core.run import (
-    RequestStatus,
-    RunHead,
-    generate_local_run_name_prefix,
-    generate_remote_run_name_prefix,
-)
+from dstack.core.run import RequestStatus, RunHead, generate_remote_run_name_prefix
 
 
 def create_run(
     storage: Storage,
-    backend_type: BackendType,
 ) -> str:
-    if backend_type is BackendType.LOCAL:
-        name = generate_local_run_name_prefix()
-    else:
-        name = generate_remote_run_name_prefix()
+    name = generate_remote_run_name_prefix()
     run_name_index = _next_run_name_index(storage, name)
     run_name = f"{name}-{run_name_index}"
     return run_name
