+++ cli/dstack/cli/commands/rm/__init__.py
@@ -1,13 +1,11 @@
 import os
-import sys
 from argparse import Namespace
 
-from rich import print
 from rich.prompt import Confirm
 
-from dstack.api.backend import list_backends
+from dstack.api.hub import HubClient
 from dstack.cli.commands import BasicCommand
-from dstack.cli.common import check_backend, check_config, check_git, check_init
+from dstack.cli.common import check_backend, check_config, check_git, check_init, console
 from dstack.cli.config import config
 from dstack.core.repo import RemoteRepo
 
@@ -46,22 +44,25 @@ class RMCommand(BasicCommand):
             repo = RemoteRepo(
                 repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd()
             )
+            hub_client = HubClient(repo=repo)
             deleted_run = False
-            for backend in list_backends(repo):
-                job_heads = backend.list_job_heads(args.run_name)
-                if job_heads:
-                    finished_job_heads = []
-                    for job_head in job_heads:
-                        if job_head.status.is_finished():
-                            finished_job_heads.append(job_head)
-                        elif args.run_name:
-                            sys.exit("The run is not finished yet. Stop the run first.")
-                    for job_head in finished_job_heads:
-                        backend.delete_job_head(job_head.job_id)
-                        deleted_run = True
+            job_heads = hub_client.list_job_heads(args.run_name)
+            if job_heads:
+                finished_job_heads = []
+                for job_head in job_heads:
+                    if job_head.status.is_finished():
+                        finished_job_heads.append(job_head)
+                    elif args.run_name:
+                        console.print("The run is not finished yet. Stop the run first.")
+                        exit(1)
+                for job_head in finished_job_heads:
+                    hub_client.delete_job_head(job_head.job_id)
+                    deleted_run = True
             if args.run_name and not deleted_run:
-                sys.exit(f"Cannot find the run '{args.run_name}'")
-            print(f"[grey58]OK[/]")
+                console.print(f"Cannot find the run '{args.run_name}'")
+                exit(1)
+            console.print(f"[grey58]OK[/]")
         else:
             if not args.run_name and not args.all:
-                sys.exit("Specify a run name or use --all to delete all runs")
+                console.print("Specify a run name or use --all to delete all runs")
+                exit(1)
