+++ cli/dstack/cli/commands/run/__init__.py
@@ -14,9 +14,8 @@ from rich.prompt import Confirm
 from websocket import WebSocketApp
 
 from dstack import providers
-from dstack.api.backend import get_backend_by_name, get_current_remote_backend, get_local_backend
-from dstack.api.run import list_runs_with_merged_backends
-from dstack.backend.base import Backend
+from dstack.api.hub import HubClient
+from dstack.api.runs import list_runs
 from dstack.backend.base.logs import fix_urls
 from dstack.cli.commands import BasicCommand
 from dstack.cli.commands.run.ssh_tunnel import allocate_local_ports, run_ssh_tunnel
@@ -35,147 +34,9 @@ from dstack.core.repo import RemoteRepo
 from dstack.core.request import RequestStatus
 from dstack.utils.workflows import load_workflows
 
-__all__ = "RunCommand"
-
 POLL_PROVISION_RATE_SECS = 3
 
 POLL_FINISHED_STATE_RATE_SECS = 1
-
-
-def _read_ssh_key_pub(key_path: str) -> str:
-    path = Path(key_path)
-    return path.with_suffix(path.suffix + ".pub").read_text().strip("\n")
-
-
-def poll_logs_ws(backend: Backend, job: Job, ports: Dict[int, int]):
-    def on_message(ws: WebSocketApp, message):
-        message = fix_urls(message, job, ports, hostname="127.0.0.1")
-        sys.stdout.buffer.write(message)
-        sys.stdout.buffer.flush()
-
-    def on_error(_: WebSocketApp, err: Exception):
-        if isinstance(err, KeyboardInterrupt):
-            run_name = job.run_name
-            if Confirm.ask(f"\n [red]Abort the run '{run_name}'?[/]"):
-                backend.stop_jobs(run_name, abort=True)
-                console.print(f"[grey58]OK[/]")
-            exit()
-        else:
-            console.print(err)
-
-    def on_open(_: WebSocketApp):
-        pass
-
-    def on_close(_: WebSocketApp, close_status_code, close_msg):
-        pass
-
-    local_ws_logs_port = ports.get(int(job.env["WS_LOGS_PORT"]), int(job.env["WS_LOGS_PORT"]))
-    url = f"ws://127.0.0.1:{local_ws_logs_port}/logsws"
-    atty = sys.stdout.isatty()
-    if atty:
-        cursor.hide()
-    _ws = websocket.WebSocketApp(
-        url,
-        on_message=on_message,
-        on_error=on_error,
-        on_open=on_open,
-        on_close=on_close,
-    )
-    _ws.run_forever()
-    if atty:
-        cursor.show()
-
-    try:
-        while True:
-            _job_head = backend.get_job(job.job_id)
-            run = backend.list_run_heads(_job_head.run_name)[0]
-            if run.status.is_finished():
-                break
-            time.sleep(POLL_FINISHED_STATE_RATE_SECS)
-    except KeyboardInterrupt:
-        if Confirm.ask(f"\n [red]Abort the run '{job.run_name}'?[/]"):
-            backend.stop_jobs(job.run_name, abort=True)
-            console.print(f"[grey58]OK[/]")
-
-
-def poll_run(
-    job_heads: List[JobHead],
-    backend: Backend,
-    ssh_key: Optional[str],
-    openssh_server: bool,
-):
-    run_name = job_heads[0].run_name
-    try:
-        console.print()
-        request_errors_printed = False
-        downloading = False
-        with Progress(
-            TextColumn("[progress.description]{task.description}"),
-            SpinnerColumn(),
-            transient=True,
-        ) as progress:
-            task = progress.add_task("Provisioning... It may take up to a minute.", total=None)
-            while True:
-                time.sleep(POLL_PROVISION_RATE_SECS)
-                run_heads = backend.list_run_heads(run_name)
-                if len(run_heads) == 0:
-                    continue
-                run = run_heads[0]
-                if run.status == JobStatus.DOWNLOADING and not downloading:
-                    progress.update(task, description="Downloading deps... It may take a while.")
-                    downloading = True
-                elif run.status not in [JobStatus.SUBMITTED, JobStatus.DOWNLOADING]:
-                    progress.update(task, total=100)
-                    break
-                if run.has_request_status([RequestStatus.TERMINATED, RequestStatus.NO_CAPACITY]):
-                    if run.has_request_status([RequestStatus.TERMINATED]):
-                        progress.update(
-                            task,
-                            description=f"[red]Request(s) terminated[/]",
-                            total=100,
-                        )
-                        break
-                    elif not request_errors_printed and run.has_request_status(
-                        [RequestStatus.NO_CAPACITY]
-                    ):
-                        progress.update(task, description=f"[dark_orange]No capacity[/]")
-                        request_errors_printed = True
-                elif request_errors_printed:
-                    progress.update(
-                        task, description="Provisioning... It may take up to a minute."
-                    )
-                    request_errors_printed = False
-
-        jobs = [backend.get_job(job_head.job_id) for job_head in job_heads]
-        ports = {}
-        if backend.name != "local":
-            console.print("Starting SSH tunnel...")
-            ports = allocate_local_ports(jobs)
-            if not run_ssh_tunnel(
-                ssh_key, jobs[0].host_name, ports
-            ):  # todo: cleanup explicitly (stop tunnel)
-                console.print("[warning]Warning: failed to start SSH tunnel[/warning] [red]✗[/]")
-        else:
-            console.print("Provisioning... It may take up to a minute. [green]✓[/]")
-        console.print()
-        console.print("[grey58]To interrupt, press Ctrl+C.[/]")
-        console.print()
-
-        if openssh_server:
-            ssh_port = jobs[0].ports[-1]
-            ssh_port = ports.get(ssh_port, ssh_port)
-            ssh_key_escaped = ssh_key.replace(" ", "\\ ")
-            console.print("To connect via SSH, use:")
-            console.print(f"  ssh -i {ssh_key_escaped} root@localhost -p {ssh_port}")
-            console.print()
-
-        run = backend.list_run_heads(run_name)[0]
-        if run.status.is_unfinished() or run.status == JobStatus.DONE:
-            poll_logs_ws(backend, jobs[0], ports)
-    except KeyboardInterrupt:
-        if Confirm.ask(f" [red]Abort the run '{run_name}'?[/]"):
-            backend.stop_jobs(run_name, abort=True)
-            console.print(f"[grey58]OK[/]")
 
 
 class RunCommand(BasicCommand):
@@ -199,13 +60,6 @@ class RunCommand(BasicCommand):
             help=workflow_help,
             choices=workflow_or_provider_names,
             nargs="?",
-        )
-        self._parser.add_argument(
-            "--remote",
-            metavar="BACKEND",
-            nargs=argparse.ZERO_OR_MORE,
-            help="",
-            type=str,
         )
         self._parser.add_argument(
             "-t",
@@ -240,21 +94,9 @@ class RunCommand(BasicCommand):
             repo = RemoteRepo(
                 repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd()
             )
-            backend = get_local_backend(repo)
-            if args.remote is not None:
-                if len(args.remote) == 0:
-                    remote_backend = get_current_remote_backend(repo)
-                    if remote_backend is None:
-                        console.print(f"No remote configured. Run `dstack config`.")
-                        exit(1)
-                else:
-                    remote_backend = get_backend_by_name(repo, args.remote[0])
-                    if remote_backend is None:
-                        console.print(f"Backend '{args.remote[0]}' is not configured")
-                        exit(1)
-                backend = remote_backend
+            hub_client = HubClient(repo=repo)
 
-            if not backend.get_repo_credentials():
+            if not hub_client.get_repo_credentials():
                 raise NotInitializedError("No credentials")
 
             if not config.repo_user_config.ssh_key_path:
@@ -263,25 +105,22 @@ class RunCommand(BasicCommand):
                 ssh_pub_key = _read_ssh_key_pub(config.repo_user_config.ssh_key_path)
 
             try:
-                run_name, jobs = backend.run_workflow(
+                run_name, jobs = hub_client.run_workflow(
                     args.workflow_or_provider,
                     ssh_pub_key=ssh_pub_key,
                     tag_name=args.tag_name,
                     args=args,
                 )
             except NameNotFoundError:
-                run_name, jobs = backend.run_provider(
+                run_name, jobs = hub_client.run_provider(
                     args.workflow_or_provider,
                     ssh_pub_key=ssh_pub_key,
                     tag_name=args.tag_name,
                     args=args,
                 )
-
-            runs_with_merged_backends = list_runs_with_merged_backends(
-                [backend], run_name=run_name
-            )
-            print_runs(runs_with_merged_backends)
-            run = runs_with_merged_backends[0][0]
+            runs = list_runs(hub_client, run_name=run_name)
+            print_runs(runs)
+            run = runs[0]
             if run.status == JobStatus.FAILED:
                 console.print("\nProvisioning failed\n")
                 exit(1)
@@ -289,9 +128,9 @@ class RunCommand(BasicCommand):
                 openssh_server = any(
                     spec.app_name == "openssh-server" for spec in jobs[0].app_specs or []
                 )
-                poll_run(
+                _poll_run(
+                    hub_client,
                     jobs,
-                    backend,
                     ssh_key=config.repo_user_config.ssh_key_path,
                     openssh_server=openssh_server,
                 )
@@ -299,3 +138,140 @@ class RunCommand(BasicCommand):
             sys.exit(
                 f"There a syntax error in one of the files inside the {os.getcwd()}/.dstack/workflows directory:\n\n{e}"
             )
+
+
+def _read_ssh_key_pub(key_path: str) -> str:
+    path = Path(key_path)
+    return path.with_suffix(path.suffix + ".pub").read_text().strip("\n")
+
+
+def _poll_run(
+    hub_client: HubClient,
+    job_heads: List[JobHead],
+    ssh_key: Optional[str],
+    openssh_server: bool,
+):
+    run_name = job_heads[0].run_name
+    try:
+        console.print()
+        request_errors_printed = False
+        downloading = False
+        with Progress(
+            TextColumn("[progress.description]{task.description}"),
+            SpinnerColumn(),
+            transient=True,
+        ) as progress:
+            task = progress.add_task("Provisioning... It may take up to a minute.", total=None)
+            while True:
+                time.sleep(POLL_PROVISION_RATE_SECS)
+                run_heads = hub_client.list_run_heads(run_name)
+                if len(run_heads) == 0:
+                    continue
+                run = run_heads[0]
+                if run.status == JobStatus.DOWNLOADING and not downloading:
+                    progress.update(task, description="Downloading deps... It may take a while.")
+                    downloading = True
+                elif run.status not in [JobStatus.SUBMITTED, JobStatus.DOWNLOADING]:
+                    progress.update(task, total=100)
+                    break
+                if run.has_request_status([RequestStatus.TERMINATED, RequestStatus.NO_CAPACITY]):
+                    if run.has_request_status([RequestStatus.TERMINATED]):
+                        progress.update(
+                            task,
+                            description=f"[red]Request(s) terminated[/]",
+                            total=100,
+                        )
+                        break
+                    elif not request_errors_printed and run.has_request_status(
+                        [RequestStatus.NO_CAPACITY]
+                    ):
+                        progress.update(task, description=f"[dark_orange]No capacity[/]")
+                        request_errors_printed = True
+                elif request_errors_printed:
+                    progress.update(
+                        task, description="Provisioning... It may take up to a minute."
+                    )
+                    request_errors_printed = False
+
+        jobs = [hub_client.get_job(job_head.job_id) for job_head in job_heads]
+        project = hub_client.get_project_info()
+        ports = {}
+        if project.backend.__root__.type != "local":
+            console.print("Starting SSH tunnel...")
+            ports = allocate_local_ports(jobs)
+            if not run_ssh_tunnel(
+                ssh_key, jobs[0].host_name, ports
+            ):  # todo: cleanup explicitly (stop tunnel)
+                console.print("[warning]Warning: failed to start SSH tunnel[/warning] [red]✗[/]")
+        else:
+            console.print("Provisioning... It may take up to a minute. [green]✓[/]")
+        console.print()
+        console.print("[grey58]To interrupt, press Ctrl+C.[/]")
+        console.print()
+
+        if openssh_server:
+            ssh_port = jobs[0].ports[-1]
+            ssh_port = ports.get(ssh_port, ssh_port)
+            ssh_key_escaped = ssh_key.replace(" ", "\\ ")
+            console.print("To connect via SSH, use:")
+            console.print(f"  ssh -i {ssh_key_escaped} root@localhost -p {ssh_port}")
+            console.print()
+
+        run = hub_client.list_run_heads(run_name)[0]
+        if run.status.is_unfinished() or run.status == JobStatus.DONE:
+            _poll_logs_ws(hub_client, jobs[0], ports)
+    except KeyboardInterrupt:
+        if Confirm.ask(f" [red]Abort the run '{run_name}'?[/]"):
+            hub_client.stop_jobs(run_name, abort=True)
+            console.print(f"[grey58]OK[/]")
+
+
+def _poll_logs_ws(hub_client: HubClient, job: Job, ports: Dict[int, int]):
+    def on_message(ws: WebSocketApp, message):
+        message = fix_urls(message, job, ports, hostname="127.0.0.1")
+        sys.stdout.buffer.write(message)
+        sys.stdout.buffer.flush()
+
+    def on_error(_: WebSocketApp, err: Exception):
+        if isinstance(err, KeyboardInterrupt):
+            run_name = job.run_name
+            if Confirm.ask(f"\n [red]Abort the run '{run_name}'?[/]"):
+                hub_client.stop_jobs(run_name, abort=True)
+                console.print(f"[grey58]OK[/]")
+            exit()
+        else:
+            console.print(err)
+
+    def on_open(_: WebSocketApp):
+        pass
+
+    def on_close(_: WebSocketApp, close_status_code, close_msg):
+        pass
+
+    local_ws_logs_port = ports.get(int(job.env["WS_LOGS_PORT"]), int(job.env["WS_LOGS_PORT"]))
+    url = f"ws://127.0.0.1:{local_ws_logs_port}/logsws"
+    atty = sys.stdout.isatty()
+    if atty:
+        cursor.hide()
+    _ws = websocket.WebSocketApp(
+        url,
+        on_message=on_message,
+        on_error=on_error,
+        on_open=on_open,
+        on_close=on_close,
+    )
+    _ws.run_forever()
+    if atty:
+        cursor.show()
+
+    try:
+        while True:
+            _job_head = hub_client.get_job(job.job_id)
+            run = hub_client.list_run_heads(_job_head.run_name)[0]
+            if run.status.is_finished():
+                break
+            time.sleep(POLL_FINISHED_STATE_RATE_SECS)
+    except KeyboardInterrupt:
+        if Confirm.ask(f"\n [red]Abort the run '{job.run_name}'?[/]"):
+            hub_client.stop_jobs(job.run_name, abort=True)
+            console.print(f"[grey58]OK[/]")
