+++ cli/dstack/cli/commands/secrets/__init__.py
@@ -1,16 +1,13 @@
 import os
-import sys
 from argparse import Namespace
 
-from rich import print
-from rich.console import Console
 from rich.prompt import Confirm, Prompt
 from rich.table import Table
 from rich_argparse import RichHelpFormatter
 
-from dstack.api.backend import list_backends
+from dstack.api.hub import HubClient
 from dstack.cli.commands import BasicCommand
-from dstack.cli.common import check_backend, check_config, check_git, check_init
+from dstack.cli.common import check_backend, check_config, check_git, check_init, console
 from dstack.cli.config import config
 from dstack.core.repo import RemoteRepo
 from dstack.core.secret import Secret
@@ -73,22 +70,22 @@ class SecretCommand(BasicCommand):
     @check_init
     def add_secret(self, args: Namespace):
         repo = RemoteRepo(repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd())
+        hub_client = HubClient(repo=repo)
         secret_value = args.secret_value or Prompt.ask("Value", password=True)
-        for backend in list_backends(repo):
-            if backend.get_secret(args.secret_name):
-                if args.yes or Confirm.ask(
-                    f"[red]The secret '{args.secret_name}' (backend: {backend.name}) already exists. "
-                    f"Do you want to override it?[/]"
-                ):
-                    backend.update_secret(
-                        Secret(secret_name=args.secret_name, secret_value=secret_value)
-                    )
-                    print(f"[grey58]OK (backend: {backend.name})[/]")
-                else:
-                    return
+        if hub_client.get_secret(args.secret_name):
+            if args.yes or Confirm.ask(
+                f"[red]The secret '{args.secret_name}' already exists. "
+                f"Do you want to override it?[/]"
+            ):
+                hub_client.update_secret(
+                    Secret(secret_name=args.secret_name, secret_value=secret_value)
+                )
+                console.print(f"[grey58]OK[/]")
             else:
-                backend.add_secret(Secret(secret_name=args.secret_name, secret_value=secret_value))
-                print(f"[grey58]OK (backend: {backend.name})[/]")
+                return
+        else:
+            hub_client.add_secret(Secret(secret_name=args.secret_name, secret_value=secret_value))
+            console.print(f"[grey58]OK[/]")
 
     @check_config
     @check_git
@@ -96,17 +93,14 @@ class SecretCommand(BasicCommand):
     @check_init
     def update_secret(self, args: Namespace):
         repo = RemoteRepo(repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd())
-        secret_value = None
-        for backend in list_backends(repo):
-            if backend.get_secret(args.secret_name):
-                if secret_value is None:
-                    secret_value = args.secret_value or Prompt.ask("Value", password=True)
-                backend.update_secret(
-                    Secret(secret_name=args.secret_name, secret_value=secret_value)
-                )
-                print(f"[grey58]OK (backend: {backend.name})[/]")
+        hub_client = HubClient(repo=repo)
+        secret_value = hub_client.get_secret(args.secret_name)
         if secret_value is None:
-            sys.exit(f"The secret '{args.secret_name}' doesn't exist")
+            console.print(f"The secret '{args.secret_name}' doesn't exist")
+            exit(1)
+        secret_value = args.secret_value or Prompt.ask("Value", password=True)
+        hub_client.update_secret(Secret(secret_name=args.secret_name, secret_value=secret_value))
+        console.print(f"[grey58]OK[/]")
 
     @check_config
     @check_git
@@ -114,37 +108,24 @@ class SecretCommand(BasicCommand):
     @check_init
     def delete_secret(self, args: Namespace):
         repo = RemoteRepo(repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd())
-        anyone = False
-        for backend in list_backends(repo):
-            secret = backend.get_secret(args.secret_name)
-            if not (secret is None) and Confirm.ask(
-                f" [red]Delete the secret '{secret.secret_name}'"
-                f"  (backend: {backend.name})?[/]"
-            ):
-                anyone = True
-                backend.delete_secret(secret.secret_name)
-                print(f"[grey58]OK[/]")
-        if not anyone:
-            sys.exit(f"The secret '{args.secret_name}' doesn't exist")
+        hub_client = HubClient(repo=repo)
+        secret = hub_client.get_secret(args.secret_name)
+        if secret is None:
+            console.print(f"The secret '{args.secret_name}' doesn't exist")
+        if Confirm.ask(f" [red]Delete the secret '{secret.secret_name}'?[/]"):
+            hub_client.delete_secret(secret.secret_name)
+            console.print(f"[grey58]OK[/]")
 
     @check_config
     @check_git
     @check_backend
     @check_init
     def _command(self, args: Namespace):
-        console = Console()
         table = Table(box=None)
-        repo = RemoteRepo(repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd())
         table.add_column("NAME", style="bold", no_wrap=True)
-        table.add_column("BACKEND", style="bold", no_wrap=True)
-        secrets = {}
-        for backend in list_backends(repo):
-            secret_names = backend.list_secret_names()
-            for secret_name in secret_names:
-                if secrets.get(secret_name) is None:
-                    secrets[secret_name] = [backend.name]
-                else:
-                    secrets[secret_name].append(backend.name)
-        for secret_name, secret_backend in secrets.items():
-            table.add_row(secret_name, ",".join(secret_backend))
+        repo = RemoteRepo(repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd())
+        hub_client = HubClient(repo=repo)
+        secret_names = hub_client.list_secret_names()
+        for secret_name in secret_names:
+            table.add_row(secret_name)
         console.print(table)
