+++ cli/dstack/cli/commands/stop/__init__.py
@@ -3,7 +3,7 @@ from argparse import Namespace
 
 from rich.prompt import Confirm
 
-from dstack.api.backend import list_backends
+from dstack.api.hub import HubClient
 from dstack.cli.commands import BasicCommand
 from dstack.cli.common import check_backend, check_config, check_git, check_init, console
 from dstack.cli.config import config
@@ -63,14 +63,12 @@ class StopCommand(BasicCommand):
             repo = RemoteRepo(
                 repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd()
             )
-            found_run = False
-            for backend in list_backends(repo):
-                job_heads = backend.list_job_heads(args.run_name)
-                found_run = len(job_heads) > 0
-                for job_head in job_heads:
-                    if job_head.status.is_unfinished():
-                        backend.stop_job(job_head.job_id, args.abort)
-            if args.run_name and not found_run:
+            hub_client = HubClient(repo=repo)
+            job_heads = hub_client.list_job_heads(args.run_name)
+            if len(job_heads) == 0:
                 console.print(f"Cannot find the run '{args.run_name}'")
                 exit(1)
+            for job_head in job_heads:
+                if job_head.status.is_unfinished():
+                    hub_client.stop_job(job_head.job_id, args.abort)
             console.print(f"[grey58]OK[/]")
