+++ cli/dstack/cli/commands/tags/__init__.py
@@ -1,18 +1,14 @@
 import os
-import sys
 from argparse import Namespace
 
 from rich import print
-from rich.console import Console
 from rich.prompt import Confirm
 from rich.table import Table
 from rich_argparse import RichHelpFormatter
 
-from dstack.api.backend import list_backends
-from dstack.api.tags import list_tag_heads_with_merged_backends
-from dstack.backend.base import BackendType
+from dstack.api.hub import HubClient
 from dstack.cli.commands import BasicCommand
-from dstack.cli.common import check_backend, check_config, check_git, check_init
+from dstack.cli.common import check_backend, check_config, check_git, check_init, console
 from dstack.cli.config import config
 from dstack.core.error import BackendError
 from dstack.core.repo import RemoteRepo
@@ -48,9 +44,6 @@ class TAGCommand(BasicCommand):
             dest="artifact_paths",
         )
         add_tags_parser.add_argument(
-            "-r", "--remote", help="Upload artifact to remote", action="store_true"
-        )
-        add_tags_parser.add_argument(
             "-y", "--yes", help="Don't ask for confirmation", action="store_true"
         )
         add_tags_parser.set_defaults(func=self.add_tag)
@@ -71,8 +64,6 @@ class TAGCommand(BasicCommand):
     @check_backend
     @check_init
     def _command(self, args: Namespace):
-        repo = RemoteRepo(repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd())
-        console = Console()
         table = Table(box=None)
         table.add_column("TAG", style="bold", no_wrap=True)
         table.add_column("CREATED", style="grey58", no_wrap=True)
@@ -80,16 +71,16 @@ class TAGCommand(BasicCommand):
         table.add_column("OWNER", style="grey58", no_wrap=True, max_width=16)
         table.add_column("BACKENDS", style="bold green", no_wrap=True)
 
-        tag_heads = list_tag_heads_with_merged_backends(list_backends(repo))
-
-        for tag_head, backends in tag_heads:
+        repo = RemoteRepo(repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd())
+        hub_client = HubClient(repo=repo)
+        tag_heads = hub_client.list_tag_heads()
+        for tag_head in tag_heads:
             created_at = pretty_date(round(tag_head.created_at / 1000))
             table.add_row(
                 tag_head.tag_name,
                 created_at,
                 tag_head.run_name,
                 tag_head.repo_user_id,
-                ", ".join(b.name for b in backends),
             )
         console.print(table)
 
@@ -98,49 +89,36 @@ class TAGCommand(BasicCommand):
     @check_backend
     @check_init
     def add_tag(self, args: Namespace):
-        if args.run_name or args.artifact_paths:
-            repo = RemoteRepo(
-                repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd()
-            )
-            added_tag = False
-            confirmed_override = False
-            for backend in list_backends(repo):
-                tag_head = backend.get_tag_head(args.tag_name)
-                if tag_head:
-                    if not args.yes and not confirmed_override:
-                        confirmed_override = Confirm.ask(
-                            f"[red]The tag '{args.tag_name}' already exists. "
-                            f"Do you want to override it?[/]"
-                        )
-                        if not confirmed_override:
-                            return
-                    backend.delete_tag_head(tag_head)
-                if args.run_name is not None:
-                    jobs_heads = backend.list_job_heads(args.run_name)
-                    if len(jobs_heads) == 0:
-                        continue
-                    try:
-                        backend.add_tag_from_run(args.tag_name, args.run_name, run_jobs=None)
-                        added_tag = True
-                    except BackendError as e:
-                        print(e)
-                        exit(1)
-                else:
-                    if not args.remote and backend.type is BackendType.REMOTE:
-                        continue
-                    if args.remote and backend.type is BackendType.LOCAL:
-                        continue
-                    try:
-                        backend.add_tag_from_local_dirs(args.tag_name, args.artifact_paths)
-                    except BackendError as e:
-                        print(e)
-                        exit(1)
-            if args.run_name is not None and not added_tag:
-                print(f"The run '{args.run_name}' doesn't exist")
+        if not args.run_name and not args.artifact_paths:
+            console.print("Specify -r RUN or -a PATH to create a tag")
+            exit(1)
+        repo = RemoteRepo(repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd())
+        hub_client = HubClient(repo=repo)
+        tag_head = hub_client.get_tag_head(args.tag_name)
+        if tag_head is not None:
+            if not args.yes and not Confirm.ask(
+                f"[red]The tag '{args.tag_name}' already exists. "
+                f"Do you want to override it?[/]"
+            ):
+                return
+            hub_client.delete_tag_head(tag_head)
+        if args.run_name is not None:
+            jobs_heads = hub_client.list_job_heads(args.run_name)
+            if len(jobs_heads) == 0:
+                console.print(f"The run '{args.run_name}' doesn't exist")
                 exit(1)
-            print(f"[grey58]OK[/]")
+            try:
+                hub_client.add_tag_from_run(args.tag_name, args.run_name, run_jobs=None)
+            except BackendError as e:
+                print(e)
+                exit(1)
         else:
-            sys.exit("Specify -r RUN or -a PATH to create a tag")
+            try:
+                hub_client.add_tag_from_local_dirs(args.tag_name, args.artifact_paths)
+            except BackendError as e:
+                print(e)
+                exit(1)
+        print(f"[grey58]OK[/]")
 
     @check_config
     @check_git
@@ -148,16 +126,12 @@ class TAGCommand(BasicCommand):
     @check_init
     def delete_tag(self, args: Namespace):
         repo = RemoteRepo(repo_ref=config.repo_user_config.repo_ref, local_repo_dir=os.getcwd())
-        tag_heads = []
-        for backend in list_backends(repo):
-            tag_head = backend.get_tag_head(args.tag_name)
-            if tag_head is not None:
-                tag_heads.append((backend, tag_head))
-
-        if len(tag_heads) == 0:
-            sys.exit(f"The tag '{args.tag_name}' doesn't exist")
+        hub_client = HubClient(repo=repo)
+        tag_head = hub_client.get_tag_head(args.tag_name)
+        if tag_head is None:
+            console.print(f"The tag '{args.tag_name}' doesn't exist")
+            exit(1)
 
         if args.yes or Confirm.ask(f" [red]Delete the tag '{args.tag_name}'?[/]"):
-            for backend, tag_head in tag_heads:
-                backend.delete_tag_head(tag_head)
+            hub_client.delete_tag_head(tag_head)
             print(f"[grey58]OK[/]")
