+++ cli/dstack/cli/common.py
@@ -7,7 +7,6 @@ from rich.console import Console
 from rich.prompt import Prompt
 from rich.table import Table
 
-from dstack.backend.base import Backend
 from dstack.core.error import BackendError, ConfigError, NotInitializedError
 from dstack.core.job import JobErrorCode, JobStatus
 from dstack.core.request import RequestStatus
@@ -74,14 +73,12 @@ def ask_choice(
                 return ask_choice(title, labels, values, selected_value, show_choices)
 
 
-def print_runs(runs_with_backends: List[Tuple[RunHead, List[Backend]]], verbose: bool = False):
-    table = generate_runs_table(runs_with_backends, verbose=verbose)
+def print_runs(runs: List[RunHead], verbose: bool = False):
+    table = generate_runs_table(runs, verbose=verbose)
     console.print(table)
 
 
-def generate_runs_table(
-    runs_with_backends: List[Tuple[RunHead, List[Backend]]], verbose: bool = False
-):
+def generate_runs_table(runs: List[RunHead], verbose: bool = False) -> Table:
     table = Table(box=None)
     table.add_column("RUN", style="bold", no_wrap=True)
     table.add_column("WORKFLOW", style="grey58", max_width=16)
@@ -89,11 +86,10 @@ def generate_runs_table(
     table.add_column("OWNER", style="grey58", no_wrap=True, max_width=16)
     table.add_column("STATUS", no_wrap=True)
     table.add_column("TAG", style="bold yellow", no_wrap=True)
-    table.add_column("BACKENDS", style="bold green", no_wrap=True)
     if verbose:
         table.add_column("ERROR", no_wrap=True)
 
-    for run, backends in runs_with_backends:
+    for run in runs:
         submitted_at = pretty_date(round(run.submitted_at / 1000))
         row = [
             _status_color(run, run.run_name, True, False),
@@ -102,7 +98,6 @@ def generate_runs_table(
             _status_color(run, run.repo_user_id or "", False, False),
             _pretty_print_status(run),
             _status_color(run, run.tag_name or "", False, False),
-            _status_color(run, ", ".join(b.name for b in backends), False, False),
         ]
         if verbose:
             row += [
