+++ cli/dstack/core/config.py
@@ -1,8 +1,5 @@
 from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import Dict, Optional
-
-from dstack.hub.models import ProjectValues
 
 
 def get_config_path():
@@ -14,36 +11,4 @@ def get_dstack_dir():
 
 
 class BackendConfig(ABC):
-    credentials: Optional[Dict] = None
-
-    @abstractmethod
-    def save(self, path: Optional[Path] = None):
-        pass
-
-    @abstractmethod
-    def load(self, path: Optional[Path] = None):
-        pass
-
-
-class Configurator(ABC):
-    NAME = ""
-
-    @property
-    def name(self):
-        return self.NAME or ""
-
-    @abstractmethod
-    def configure_cli(self):
-        pass
-
-    @abstractmethod
-    def configure_hub(self, config_data: Dict) -> ProjectValues:
-        pass
-
-    @abstractmethod
-    def get_config_from_hub_config_data(self, config_data: Dict, auth_data: Dict) -> BackendConfig:
-        pass
-
-    @abstractmethod
-    def register_parser(self, parser):
-        pass
+    pass
